Move WebPathResolver from Pimcore\Service to Pimcore\HttpKernel
