+++ pimcore/lib/Pimcore/Document/Tag/TagHandler.php
@@ -20,10 +20,10 @@ use Pimcore\Extension\Document\Areabrick\Exception\ConfigurationException;
 use Pimcore\Extension\Document\Areabrick\TemplateAreabrickInterface;
 use Pimcore\Http\RequestHelper;
 use Pimcore\HttpKernel\BundleLocator\BundleLocatorInterface;
+use Pimcore\HttpKernel\WebPathResolver;
 use Pimcore\Model\Document\PageSnippet;
 use Pimcore\Model\Document\Tag;
 use Pimcore\Model\Document\Tag\Area\Info;
-use Pimcore\Service\WebPathResolver;
 use Pimcore\Templating\Model\ViewModel;
 use Pimcore\Templating\Model\ViewModelInterface;
 use Pimcore\Templating\Renderer\ActionRenderer;
